requirements.txt
python-telegram-bot[ext]~=21.0
python-docx~=1.1
mistune~=3.0
python-dotenv~=1.0
pyyaml~=6.0
